requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
langchain==0.3.27
langgraph==0.6.6
langchain-ollama==0.3.6
langchain-mcp-adapters==0.1.9
pydantic-settings==2.10.1
